back ToTop function
changed function for backToTop for mobile also
