+++ explore/index.html
@@ -33,7 +33,6 @@
     />
 
     <script type="module" src="/src/js/pages/explore/index.js"></script>
-    <script defer src="backToTop.js"></script>
   </head>
 
   <body>
@@ -440,16 +439,16 @@
     </div>
 
     <a
-      href="#logo"
+      href="#"
       id="backButton"
-      class="fixed bottom-24 right-1/2 xl:right-60 2xl:right-1/4 items-center gap-5 hidden"
+      class="fixed bottom-24 right-2 md:right-4 xl:right-60 2xl:right-1/4 items-center gap-5 hidden"
     >
       <span
         class="rounded-full w-12 h-12 bg-light flex items-center justify-center"
       >
         <i class="fa-solid fa-chevron-up"></i>
       </span>
-      <span>Back to top</span>
+      <span class="hidden xl:inline-block">Back to top</span>
     </a>
     <!-- Modal container -->
